refactor: centralize Slack channel permission deletion
