+++ services/console/app/views/console/principal.html.erb
@@ -106,7 +106,7 @@
                       <%= button_tag type: "submit",
                                      name: "_method",
                                      value: "delete",
-                                     formaction: console_principal_slack_channel_permission_path(@principal.oid, permission.id),
+                                     formaction: console_slack_channel_permission_path(permission.id),
                                      formmethod: "post",
                                      class: "grid size-8 place-items-center rounded text-zinc-500 transition-colors hover:bg-red-500/10 hover:text-red-300 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-red-400",
                                      aria: { label: "Delete #{permission.channel_id} Slack channel permission" },
